Make work page the home page
